move function matches to cfg
