+++ pkg/domains/domains_test.go
@@ -37,16 +37,16 @@ func TestIsUnderManagement(t *testing.T) {
 
 func TestMatches(t *testing.T) {
 	// Full email should not be accepted
-	assert.Equal(t, "", Matches("test@vouch.github.io"))
+	assert.Equal(t, "", cfg.Matches("test@vouch.github.io"))
 
-	assert.Equal(t, "vouch.github.io", Matches("vouch.github.io"))
-	assert.Equal(t, "vouch.github.io", Matches("sub.vouch.github.io"))
-	assert.Equal(t, "", Matches("a-different-vouch.github.io"))
+	assert.Equal(t, "vouch.github.io", cfg.Matches("vouch.github.io"))
+	assert.Equal(t, "vouch.github.io", cfg.Matches("sub.vouch.github.io"))
+	assert.Equal(t, "", cfg.Matches("a-different-vouch.github.io"))
 
-	assert.Equal(t, "", Matches("mydomain.com"))
+	assert.Equal(t, "", cfg.Matches("mydomain.com"))
 
-	assert.Equal(t, "test.mydomain.com", Matches("test.mydomain.com"))
-	assert.Equal(t, "sub.test.mydomain.com", Matches("sub.test.mydomain.com"))
-	assert.Equal(t, "sub.test.mydomain.com", Matches("subsub.sub.test.mydomain.com"))
-	assert.Equal(t, "test.mydomain.com", Matches("other.test.mydomain.com"))
+	assert.Equal(t, "test.mydomain.com", cfg.Matches("test.mydomain.com"))
+	assert.Equal(t, "sub.test.mydomain.com", cfg.Matches("sub.test.mydomain.com"))
+	assert.Equal(t, "sub.test.mydomain.com", cfg.Matches("subsub.sub.test.mydomain.com"))
+	assert.Equal(t, "test.mydomain.com", cfg.Matches("other.test.mydomain.com"))
 }
